Start on function to check if king in check
added comments and started on function to check of the king is in check. however, this has not been implemented yet since it does not work. if the king is in check, the function returns no possible moves to get out of check, even if it is possible to do so.